Backward linkages

Backward linkages as defined by the Eurostat Manual of Supply, Use and Input-Output Tables (see p506-507.)

backward_linkages(Im)

Arguments

Im
A Leontieff inverse matrix created by the leontieff_inverse_create function.

Value

The vector of industry (product) backward linkages in a wide data.frame class, following the column names of the Leontieff inverse matrix.

See also

Examples

de_coeff <- input_coefficient_matrix_create( iotable_get(), 
                                             digits = 4 )

#> Warning: `funs()` was deprecated in dplyr 0.8.0.
#> Please use a list of either functions or lambdas: 
#> 
#>   # Simple named list: 
#>   list(mean = mean, median = median)
#> 
#>   # Auto named with `tibble::lst()`: 
#>   tibble::lst(mean, median)
#> 
#>   # Using lambdas
#>   list(~ mean(., trim = .2), ~ median(., na.rm = TRUE))

I <- leontieff_inverse_create (de_coeff)
backward_linkages (I)

#>        iotables_row agriculture_group industry_group construction trade_group
#> 1 backward linkages          1.704922       1.841421     1.813747    1.603712
#>   business_services_group other_services_group
#> 1                1.594974             1.378375