Create a coefficient matrix

Create a coefficient matrix from a Symmetric Input-Output Table. The coefficient matrix is related by default to output, but you can change this to total supply or other total aggregate if it exists in your table.

coefficient_matrix_create(siot, total = "output", digits = NULL)

Arguments

siot
A symmetric input-output table retrieved by the iotable_get function.

total
Usually an output vector with a key column, defaults to "output" which equals "P1" or "output_bp". You can use other rows for comparison, for example "TS_BP" if it exists in the matrix.

digits
An integer showing the precision of the technology matrix in digits. Default is NULL when no rounding is applied.

Value

A data.frame that contains the matrix of siot devided by total with a key column.

References

United Kingdom Input-Output Analytical Tables 2010 for explanation on the use of the Coefficient matrix.

Examples

coefficient_matrix_create ( siot = iotable_get ( source = "germany_1990"),
                            total = "output",
                            digits = 4 )

#>                   iotables_row agriculture_group manufacturing_group
#> 1            agriculture_group            0.0258              0.0236
#> 2          manufacturing_group            0.1806              0.2822
#> 3           construction_group            0.0097              0.0068
#> 4                  trade_group            0.0811              0.0674
#> 5      business_services_group            0.0828              0.0890
#> 6         other_services_group            0.0353              0.0139
#> 7                        total            0.4153              0.4829
#> 8        import_goods_services            0.0667              0.1452
#> 9           net_tax_production            0.0247              0.0060
#> 10 intermediate_consumption_pp            0.5066              0.6341
#> 11      compensation_employees            0.2137              0.2746
#> 12    net_tax_production_other           -0.0458              0.0013
#> 13   consumption_fixed_capital            0.1793              0.0591
#> 14       operating_surplus_net            0.1463              0.0309
#> 15                      gva_bp            0.4934              0.3659
#> 16                   output_bp            1.0000              1.0000
#> 17          wage_salary_earner            0.0110              0.0074
#> 18               self_employed            0.0140              0.0003
#> 19            employment_total            0.0250              0.0078
#>    construction_group trade_group business_services_group other_services_group
#> 1              0.0000      0.0011                  0.0010               0.0015
#> 2              0.2613      0.0761                  0.0173               0.0597
#> 3              0.0158      0.0098                  0.0339               0.0180
#> 4              0.0578      0.1378                  0.0156               0.0413
#> 5              0.1263      0.1218                  0.2790               0.0672
#> 6              0.0071      0.0208                  0.0217               0.0434
#> 7              0.4683      0.3673                  0.3686               0.2310
#> 8              0.0547      0.0406                  0.0193               0.0271
#> 9              0.0063      0.0155                  0.0122               0.0247
#> 10             0.5292      0.4234                  0.4001               0.2828
#> 11             0.3209      0.3971                  0.1802               0.5364
#> 12             0.0039      0.0051                  0.0086              -0.0169
#> 13             0.0239      0.0761                  0.1424               0.0968
#> 14             0.1221      0.0983                  0.2687               0.1010
#> 15             0.4708      0.5766                  0.5999               0.7172
#> 16             1.0000      1.0000                  1.0000               1.0000
#> 17             0.0118      0.0148                  0.0053               0.0188
#> 18             0.0014      0.0024                  0.0009               0.0013
#> 19             0.0132      0.0171                  0.0061               0.0201