Input-output table for Croatia, 2010.

1700 - Symmetric input-output table at basic prices (product x product) In thousand kunas (T_NAC)

data(croatia_2010_1700)

Format

A data frame with 13 variables.

t_rows2

Technology codes in row names, following the Eurostat convention.

t_rows2_lab

Longer labels for t_rows2

t_cols2

Technology codes in column names, following the Eurostat convention.

t_cols2_lab

Longer labels for t_cols2

iotables_col

The standardized iotables column labelling for easier reading.

col_order

The column ordering to keep the matrix legible.

row_order

The row ordering to keep the matrix legible.

iotables_row

The standardized iotables row labelling for easier reading.

unit

Different from Eurostat tables, in thousand national currency units.

geo

ISO / Eurostat country code for Croatia

geo_lab

ISO / Eurostat country name, Croatia.

time

Date of the SIOT

values

The actual values of the table in thousand kunas

Source

Eurostat Manual of Supply, Use and Input-Output Tables Updated 17 December 2015.