Forward linkages

Forward linkages as defined by the Eurostat Manual of Supply, Use and Input-Output Tables (see p506-507.)

forward_linkages(output_coefficient_matrix, digits = NULL)

Arguments

output_coefficient_matrix
An output coefficient matrix created with the output_coefficient_matrix_create function.

digits
Number of decimals for rounding, defaults to NULL.

Value

The vector of industry (product) forward linkages in a long-form data.frame, containing the metadata column of the the row names from the output_coefficient_matrix.

See also

Examples

data_table = iotable_get()

de_out <- output_coefficient_matrix_create (
 data_table, "tfu", digits = 4
 )

forward_linkages ( output_coefficient_matrix = de_out,
                   digits = 4 )

#>              iotables_row forward_linkages
#> 1       agriculture_group           2.1128
#> 2          industry_group           1.6909
#> 3            construction           1.3559
#> 4             trade_group           1.5850
#> 5 business_services_group           2.1040
#> 6    other_services_group           1.2105