Simple input-output table for Germany, 1990.

For testing purposes a well documented example data set is used from the Eurostat manual. The table in the Eurostat manual is brought to the format used by the Eurostat database. It is a small dataset for examples, but it is also instructive to understand how Eurostat stores the highly structured SIOTs in long-form tidy datasets. The labels were slightly alterred to reflect the transition from the vocabulary of ESA95 to ESA2010 since the publication of the Manual.

Usage

data(germany_1990)

Format

A data frame with 228 observations and 10 variables.

prod_na

Technology codes in row names, following the Eurostat convention.

prod_na_lab

Longer labels for t_rows2

induse

Column labels, following the Eurostat convention with differences.

iotables_row

Row labels, i.e. to be used in key column, for iotables package abbreviations

iotables_col

Column labels for iotables package abbreviations

values

The actual values of the table in million euros

unit

MIO_EUR, the same as Eurostat

unit_lab

Million euros. Eurostat usually has euro and national currency unit values, too.

geo

ISO / Eurostat country code for Germany, i.e. DE

geo_lab

ISO / Eurostat country name, Germany

time

Date of the SIOT

Source

Eurostat Manual of Supply, Use and Input-Output Tables p 492

See also