Get the gross value added (GVA) vector

Get the gross value added (GVA) vector

gva_get(labelled_io_table = NULL, source = "germany_1990",
  geo = "DE", year = 1990, unit = "MIO_EUR", households = FALSE,
  labelling = "iotables", stk_flow = "DOM", keep_total = FALSE)

Arguments

labelled_io_table
If you have created the IO table earlier with iotable_get, it is faster to work with the data in the memory. Defaults to NULL when the data will be retrieved from the hard disk or from the Eurostat website invoking iotables_download and iotable_get

source
A data source, for example "naio_10_cp1700". Possible codes are "naio_10_cp1700", "naio_10_cp1750", "naio_10_pyp1700", "naio_10_pyp1750", "naio_10_cp1620", "naio_10_pyp1620", "naio_10_cp1630", "naio_10_pyp1630", "croatia_2010_1700", "croatia_2010_1800", "croatia_2010_1900". For further information consult the Eurostat Symmetric Input-Output Tables page.

geo
A country code or a country name. For example, SK for Slovakia.

year
A numeric variable containing the year. Defaults to 2010, because this year has the most data.

unit
A character string containing the currency unit, defaults to MIO_NAC (million national currency unit). The alternative is MIO_EUR.

households
If you need to make household demand endogenous, or "close the households off", TRUE selects wages and final household consumption. This is needed for induced-effects calculations.

labelling
Defaults to iotables which gives standard row and column names regardless of the source of the table, or if it is a product x product, industry x industry or product x industry table. The alternative is short which is the original short row or column code of Eurostat or OECD.

stk_flow
Defaults to DOM, alternative IMP.

keep_total
Logical variable. Defaults to FALSE and removes the totaling row and column from the matrix.

Examples

gva_get (  source = "germany_1990", geo = "DE",
              year = 1990, unit = "MIO_EUR",
              stk_flow = "DOM", labelling = "iotables")

#>    iotables_row agriculture_group manufacturing_group construction_group
#> 15       gva_bp             21664              395022             115624
#>    trade_group business_services_group other_services_group
#> 15      311407                  415426               365017
# NOT RUN {
GVA <-  iotables_download ( "naio_10_cp1700" ) %>%
 iotable_get (labelled_io_data = ., geo = "CZ",
              source = "naio_10_cp1700",
              year = 2015, unit = "MIO_NAC",
              labelling = "short") %>%
 gva_get ()
# }