Create an input coefficient matrix

Create an input coefficient matrix from the input flow matrix and the output vector. The two input vectors must have consistent labelling, i.e the same column names must be found in the use table (input flow) and the output vector.

Usage

input_coefficient_matrix_create(data_table, households = FALSE, digits = NULL)

Arguments

data_table

A symmetric input-output table, a use table, a margins or tax table retrieved by the iotable_get function.

households

Defaults to NULL. Household column can be added with TRUE.

digits

An integer showing the precision of the technology matrix in digits. Default is NULL when no rounding is applied.

Value

A data frame that contains the matrix of first quadrant of the use table as input_flow divided by output supported by a key column of product or industries, with a key column. Optionally the results are rounded to given digits. An input coefficient matrix of data.frame class. The column names are ordered, and the row names are in the first, auxiliary metadata column.

Details

The terminology follows the Eurostat Manual of Supply, Use and Input-Output Tables. Input-Output Multipliers Specification Sheet and Supporting Material, Spicosa Project Report, which cannot be linked due to a malformatted url, but can be found with a search engine. this matrix is called 'technological coefficients'. The results of the function are tested on both sources.This is a wrapper function around coefficient_matrix_create.

Examples

input_coefficient_matrix_create ( 
                           iotable_get(), 
                           digits = 4 )
#>              iotables_row agriculture_group industry_group construction
#> 1       agriculture_group            0.0258         0.0236       0.0000
#> 2          industry_group            0.1806         0.2822       0.2613
#> 3            construction            0.0097         0.0068       0.0158
#> 4             trade_group            0.0811         0.0674       0.0578
#> 5 business_services_group            0.0828         0.0890       0.1263
#> 6    other_services_group            0.0353         0.0139       0.0071
#>   trade_group business_services_group other_services_group
#> 1      0.0011                  0.0010               0.0015
#> 2      0.0761                  0.0173               0.0597
#> 3      0.0098                  0.0339               0.0180
#> 4      0.1378                  0.0156               0.0413
#> 5      0.1218                  0.2790               0.0672
#> 6      0.0208                  0.0217               0.0434
                           
#This is a wrapper function and equivalent to                           

coefficient_matrix_create( iotable_get(), 
                           total = "total", 
                           return = "products")
#>              iotables_row agriculture_group industry_group construction
#> 1       agriculture_group        0.06202358     0.04888568 8.695123e-06
#> 2          industry_group        0.43487798     0.58437193 5.579400e-01
#> 3            construction        0.02336167     0.01407094 3.369360e-02
#> 4             trade_group        0.19517412     0.13951414 1.233838e-01
#> 5 business_services_group        0.19945160     0.18440531 2.697836e-01
#> 6    other_services_group        0.08511105     0.02875200 1.519038e-02
#> 7                   total        1.00000000     1.00000000 1.000000e+00
#>   trade_group business_services_group other_services_group
#> 1 0.003060031             0.002781946          0.006480804
#> 2 0.207104112             0.046944365          0.258211570
#> 3 0.026698393             0.091910022          0.077863206
#> 4 0.375063015             0.042454068          0.178672881
#> 5 0.331486560             0.756908827          0.291066356
#> 6 0.056587889             0.059000772          0.187705183
#> 7 1.000000000             1.000000000          1.000000000