Create input indicator(s)

The function creates the input indicators from the inputs and the outputs.

input_indicator_create(
  data_table,
  input_vector = c("gva_bp", "net_tax_production"),
  digits = NULL,
  households = FALSE,
  indicator_names = NULL
)

Arguments

data_table
A symmetric input-output table, a use table, a margins or tax table retrieved by the iotable_get function.

input_vector
The name of inputs for which you want to create the indicators. They must be found in the data_table.

digits
Rounding digits, if omitted, no rounding takes place.

households
If the households column should be added, defaults to FALSE.

indicator_names
The names of new indicators. Defaults to NULL when the names in the key colum of input_matrix will be used to create the indicator names.

Value

A tibble (data frame) containing input_matrix devided by the output_vector with a key column for products or industries.

See also

Examples

 
input_indicator_create( data_table = iotable_get(), 
                        input_vector = c("gva", "compensation_employees"),
                        digits = 4, 
                        indicator_names = c("GVA indicator", "Income indicator"))

#>        iotables_row agriculture_group industry_group construction trade_group
#> 10    GVA indicator            0.2137         0.2746       0.3209      0.3971
#> 14 Income indicator            0.4934         0.3659       0.4708      0.5766
#>    business_services_group other_services_group
#> 10                  0.1802               0.5364
#> 14                  0.5999               0.7172