Download input-output tables

This function downloads standard input-output table files. Currently only Eurostat files are supported. You are not likely to use this function, because iotable_get will call this function if necessary and properly filter out an input-output table.

Usage

iotables_download(
  source = "naio_10_cp1700",
  data_directory = NULL,
  force_download = FALSE
)

Arguments

source

See the available list of sources above in the Description.

data_directory

Defaults to NULL when the files will be temporarily stored in the path retrieved by tempdir. If it is a different valid directory, it will try to save the pre-processed data file here with labelling.

force_download

Defaults to FALSE which will use the existing downloaded file in the data_directory or the temporary directory, if it exists. TRUE will try to download the file from the Eurostat warehouse.

Value

A nested data frame. Each input-output table is in a separate row of the nested output, where all the metadata are in columns, and the actual, tidy, ordered input-output table is in the data data column. The data is saved into the actual tempdir(), too.

Details

The data is downloaded in the tempdir()under the name the statistical product as an rds file. (For example: naio_10_cp1750.rds) The temporary directory is emptied at every normal R session exit.

To save the file for further use (which is necessary in analytical work because download times are long) set the download_directory [see parameters]. The function will make a copy of the rds file in this directory.

naio_10_cp1700Symmetric input-output table at basic prices (product by product)

naio_10_pyp1700Symmetric input-output table at basic prices (product by product) (previous years prices)

naio_10_cp1750Symmetric input-output table at basic prices (industry by industry)

naio_10_pyp1750Symmetric input-output table at basic prices (industry by industry) (previous years prices)

naio_10_cp15Supply table at basic prices incl. transformation into purchasers' prices

naio_10_cp16Use table at purchasers' prices

naio_10_cp1610Use table at basic prices

naio_10_pyp1610Use table at basic prices (previous years prices) (naio_10_pyp1610)

naio_10_cp1620Table of trade and transport margins at basic prices

naio_10_pyp1620Table of trade and transport margins at previous years' prices

naio_10_cp1630Table of taxes less subsidies on products at basic prices

naio_10_pyp1630Table of taxes less subsidies on products at previous years' prices

uk_2010_siotUnited Kingdom Input-Output Analytical Tables data

See also

Examples

# \donttest{
 io_tables <- iotables_download(source = "naio_10_pyp1750")
#> Table naio_10_pyp1750 cached at C:\Users\DANIEL~1\AppData\Local\Temp\RtmpWMjwEy/eurostat/naio_10_pyp1750_date_code_FF.rds
#> Saving 96 input-output tables into the temporary directory.
#> Saved the raw data of this table type in temporary directory C:\Users\DANIEL~1\AppData\Local\Temp\RtmpWMjwEy/naio_10_pyp1750_processed.rds.
 # }