Get Metadata from Nested iotables File

Remove the data column and return only the metadata information of input-output (or related tables) from a source. If dat is not inputed as a nested data frame created by iotables_download, validate the source input parameter and try to load the table from the current sessions' temporary directory.

naio_10_cp1700 Symmetric input-output table at basic prices (product by product)

naio_10_pyp1700 Symmetric input-output table at basic prices (product by product) (previous years prices)

naio_10_cp1750 Symmetric input-output table at basic prices (industry by industry)

naio_10_pyp1750 Symmetric input-output table at basic prices (industry by industry) (previous years prices)

naio_10_cp15 Supply table at basic prices incl. transformation into purchasers' prices

naio_10_cp16 Use table at purchasers' prices

naio_10_cp1610 Use table at basic prices

naio_10_pyp1610 Use table at basic prices (previous years prices) (naio_10_pyp1610)

naio_10_cp1620 Table of trade and transport margins at basic prices

naio_10_pyp1620 Table of trade and transport margins at previous years' prices

naio_10_cp1630 Table of taxes less subsidies on products at basic prices

naio_10_pyp1630Table of taxes less subsidies on products at previous years' prices

uk_2010_siotUnited Kingdom Input-Output Analytical Tables data

iotables_metadata_get(dat = NULL, source = "naio_10_cp1700")

Arguments

dat
A nested data file created by iotables_download. Defaults to NULL in which case an attempt is made to find and read in the nested data from the current R sessions' temporary directory.

source
See the available list of sources above in the Description.

Value

A data frame, which contains the metadata of all available input-output tables from a specific source.

See also

Examples

iotables_metadata_get ( source = "naio_10_cp1700" )

#> The temporary file for source='naio_10_cp1700' is not found in
#> tempdir='C:\Users\DANIEL~1\AppData\Local\Temp\RtmpYLojZC'