Metadata

An arrangement of the Eurostat national accounts vocabulary, used to correctly order wide format rows and columns from bulk long-form tables.

data(metadata)

Format

A data frame with 10 variables. #'

variable

Eurostat vocabulary source, i.e. t_rows, t_cols, prod_na, induse

group

Informal labelling for macroeconomic groups

code

Eurostat labels

label

Eurostat label descriptions

quadrant

Where to place the data from a long-form raw data file

account_group

Different from Eurostat tables, in thousand national currency units.

digit_1

third digit for ordering

digit_2

fourth digit for ordering

numeric_label

ordering from quadrant, account_group, digit_1, digit_2

iotables_label

Custom, machine_readable snake format variable names