Create an output coefficient matrix

Create an output coefficient matrix from the input flow matrix or a symmetric input-output table. If there are zero values in present, they will be changed to 0.000001 and you will get a warning. Some analytical equations cannot be solved with zero elements. You either have faulty input data, or you have to use some sort of data modification to carry on your analysis.

output_coefficient_matrix_create(io_table, type = "product",
  digits = NULL)

Arguments

io_table
An input flow matrix created with the use_table_get function which contains the 'total' column is sufficient if type=products is used. In case you use type="final_demand" you need to input a full iotable, create by the iotable_get, because you will need the final demand column.

type
The type=products (default) returns the output coefficients for products (intermediates) while the final_demand returns output coefficients for final demand. See Eurostat Manual, p495 and p507.

digits
An integer showing the precision of the technology matrix in digits. Default is NULL when no rounding is applied.

Examples

#You need a table that has a total column and either the total 
#intermediate use or final use
#This is usually the case with Eurostat tables, but with the Germany data
#file total must be added.

io_table <- iotable_get ()
io_table <- io_table [1:which(tolower(io_table[,1]) =="total" ), ]
output_bp <- dplyr::select ( io_table, output_bp )
io_table <- io_table [, 1:7]
io_table$total <- rowSums(io_table[, 2:7])
io_table <- cbind (io_table, output_bp)

output_coefficient_matrix_create ( io_table = io_table,
                                    type = 'final_demand',
                                    digits = 4 )

#>              iotables_row agriculture_group manufacturing_group
#> 1       agriculture_group            0.0258              0.5803
#> 2     manufacturing_group            0.0073              0.2822
#> 3      construction_group            0.0017              0.0299
#> 4             trade_group            0.0066              0.1346
#> 5 business_services_group            0.0053              0.1388
#> 6    other_services_group            0.0030              0.0294
#>   construction_group trade_group business_services_group other_services_group
#> 1             0.0000      0.0138                  0.0162               0.0174
#> 2             0.0594      0.0381                  0.0111               0.0281
#> 3             0.0158      0.0216                  0.0955               0.0373
#> 4             0.0263      0.1378                  0.0201               0.0389
#> 5             0.0448      0.0950                  0.2790               0.0494
#> 6             0.0034      0.0221                  0.0296               0.0434