Create a coefficient matrix

Create a coefficient matrix from a Symmetric Input-Output Table. The coefficient matrix is related by default to output, but you can change this to total supply or other total aggregate if it exists in your table.

coefficient_matrix_create(
  data_table,
  total = "output",
  digits = NULL,
  remove_empty = TRUE,
  households = FALSE,
  return_part = NULL
)

Arguments

data_table
A symmetric input-output table, a use table, a margins or tax table retrieved by the iotable_get function.

total
Usually an output vector with a key column, defaults to "output" which equals "P1" or "output_bp". You can use other rows for comparison, for example "TS_BP" if it exists in the matrix.

digits
An integer showing the precision of the technology matrix in digits. Default is NULL when no rounding is applied.

remove_empty
Defaults to TRUE. If you want to keep empty primary input rows, choose FALSE. Empty product/industry rows are always removed to avoid division by zero error in the analytical functions.

households
Defaults to NULL. Household column can be added with TRUE.

return_part
Defaults to NULL. You can choose "product" or "industry" to return an input coefficient matrix or "primary_inputs" to get only the total intermediate use and proportional primary inputs.

Value

A data.frame that contains the matrix of data_table divided by total with a key column. Optionally the results are rounded to given digits.

References

See United Kingdom Input-Output Analytical Tables 2010 for explanation on the use of the Coefficient matrix.

Examples

coefficient_matrix_create ( data_table = iotable_get ( source = "germany_1990"),
                            total = "output",
                            digits = 4 )

#>                 iotables_row agriculture_group industry_group construction
#> 1          agriculture_group            0.0258         0.0236       0.0000
#> 2             industry_group            0.1806         0.2822       0.2613
#> 3               construction            0.0097         0.0068       0.0158
#> 4                trade_group            0.0811         0.0674       0.0578
#> 5    business_services_group            0.0828         0.0890       0.1263
#> 6       other_services_group            0.0353         0.0139       0.0071
#> 7                      total            0.4153         0.4829       0.4683
#> 8                    imports            0.0667         0.1452       0.0547
#> 9   intermediate_consumption            0.5066         0.6341       0.5292
#> 10    compensation_employees            0.2137         0.2746       0.3209
#> 11        net_tax_production           -0.0458         0.0013       0.0039
#> 12 consumption_fixed_capital            0.1793         0.0591       0.0239
#> 13       os_mixed_income_net            0.1463         0.0309       0.1221
#> 14                       gva            0.4934         0.3659       0.4708
#> 15                    output            1.0000         1.0000       1.0000
#> 16          net_tax_products            0.0247         0.0060       0.0063
#> 17    employment_wage_salary            0.0110         0.0074       0.0118
#> 18  employment_self_employed            0.0140         0.0003       0.0014
#> 19 employment_domestic_total            0.0250         0.0078       0.0132
#>    trade_group business_services_group other_services_group
#> 1       0.0011                  0.0010               0.0015
#> 2       0.0761                  0.0173               0.0597
#> 3       0.0098                  0.0339               0.0180
#> 4       0.1378                  0.0156               0.0413
#> 5       0.1218                  0.2790               0.0672
#> 6       0.0208                  0.0217               0.0434
#> 7       0.3673                  0.3686               0.2310
#> 8       0.0406                  0.0193               0.0271
#> 9       0.4234                  0.4001               0.2828
#> 10      0.3971                  0.1802               0.5364
#> 11      0.0051                  0.0086              -0.0169
#> 12      0.0761                  0.1424               0.0968
#> 13      0.0983                  0.2687               0.1010
#> 14      0.5766                  0.5999               0.7172
#> 15      1.0000                  1.0000               1.0000
#> 16      0.0155                  0.0122               0.0247
#> 17      0.0148                  0.0053               0.0188
#> 18      0.0024                  0.0009               0.0013
#> 19      0.0171                  0.0061               0.0201