Get employment data

Download the employment data for a country and arrange it to the 64x64 SIOTs.

Usage

employment_get(
  geo = "CZ",
  year = "2010",
  sex = "Total",
  age = "Y_GE15",
  labelling = "iotables",
  data_directory = NULL,
  force_download = TRUE
)

Source

Eurostat statistic Employment by sex, age and detailed economic activity (from 2008 onwards, NACE Rev. 2 two digit level) - 1 000

Arguments

geo

The country code.

year

The year. The average employment will be created for the given year, starting with 2008, when the NACE Rev 2 was introduced in employment statistics.

sex

Defaults to "Total". Enter "Females" or "F" for female employment, "Males" or "M" for male employment.

age

Defaults to "Y_GE15", which is the Eurostat code for employment in all age groups starting from 15-years-old. Any Eurostat code can be used as a parameter.

labelling

Either "iotables" or the applicable short code, for product x product SIOTs "prod_na" and in the case of industry x industry SIOTs "induse".

data_directory

Defaults to NULL, if a valid directory, it will try to save the pre-processed data file here with labelling.

force_download

Defaults to TRUE. If FALSE it will use the existing downloaded file in the data_directory or the temporary directory, if it exists.

Value

A data.frame with auxiliary metadata to conform the symmetric input-output tables.

Details

Currently works only with product x product tables.

See also

Examples

if (FALSE) {
 io_tables <- get_employment ( 
               geo = "CZ", 
               year = "2010",
               sex = "Total", 
               age = "Y_GE15",
               data_directory = NULL,
               force_download = TRUE
               )
 }