Get An Input-Output Table Fom Bulk File

This function is used to filter out a single input-output table from a database, for example a raw file downloaded from the Eurostat website. It provides some functionality to avoid some pitfalls.

iotable_get(
  labelled_io_data = NULL,
  source = "germany_1990",
  geo = "DE",
  year = 1990,
  unit = "MIO_EUR",
  stk_flow = "DOM",
  labelling = "iotables",
  data_directory = NULL,
  force_download = TRUE
)

Arguments

labelled_io_data
If you have downloaded a bulk data file with iotables_download, it is faster to work with the data in the memory. Defaults to NULL when the data will be retrieved from the hard disk or from the Eurostat website invoking the same function.

source
A data source, for example naio_10_cp1700.

naio_10_cp1700 Symmetric input-output table at basic prices (product by product)

naio_10_pyp1700 Symmetric input-output table at basic prices (product by product) (previous years prices)

naio_10_cp1750 Symmetric input-output table at basic prices (industry by industry)

naio_10_pyp1750 Symmetric input-output table at basic prices (industry by industry) (previous years prices)

naio_10_cp15 Supply table at basic prices incl. transformation into purchasers' prices

naio_10_cp16 Use table at purchasers' prices

naio_10_cp1610 Use table at basic prices

naio_10_pyp1610 Use table at basic prices (previous years prices) (naio_10_pyp1610)

naio_10_cp1620 Table of trade and transport margins at basic prices

naio_10_pyp1620 Table of trade and transport margins at previous years' prices

naio_10_cp1630 Table of taxes less subsidies on products at basic prices

naio_10_pyp1630Table of taxes less subsidies on products at previous years' prices

For further information consult the Eurostat Symmetric Input-Output Tables page.

geo
A country code or a country name. For example, SK or as Slovakia.

year
A numeric variable containing the year. Defaults to 2010, because this year has the most data.

unit
A character string containing the currency unit, defaults to MIO_NAC (million national currency unit). The alternative is MIO_EUR.

stk_flow
Defaults to DOM as domestic output, alternative IMP for imports and TOTAL for total output. For source = 'naio_10_cp1620' and trade and transport margins and source = 'naio_10_cp1630' taxes less subsidies only TOTAL is not used.

labelling
Defaults to iotables which gives standard row and column names regardless of the source of the table, or if it is a product x product, industry x industry or product x industry table. The alternative is short or eurostat which is the original short row or column code of Eurostat or OECD.

data_directory
Defaults to NULL, if a valid directory, it will try to save the pre-processed data file here with labelling.

force_download
Defaults to TRUE. If FALSE it will use the existing downloaded file in the data_directory or the temporary directory, if it exists. Will force download only in a new session.

Value

A wide format data.frame with a well-ordered input-output table. The bulk data files on the Eurostat website are in a long form and they are not correctly ordered for further matrix equations.

Details

Unless you want to work with bulk data files, you should not invoke iotables_download directly, rather via this function, if and when it is necessary.

See also

Examples

germany_table <- iotable_get( source = "germany_1990",
                 geo = 'DE', year = 1990, unit = "MIO_EUR",
                 labelling  = "iotables" )