Download input-output tables

This function downloads standard input-output table files. Currently only Eurostat files are supported. You are not likely to use this function, because iotable_get will call this function if necessary and properly filter out an input-output table. The only parameter is the Eurostat code of the table. The data is downloaded in the tempdir()under the name the statistical product as an rds file. (For example: naio_10_cp1750.rds) The temporary directory is emptied at every normal R session exit. To save the file for further use (which is necessary in analytical work because download times are long) set the download_directory [see parameters]. The function will make a copy of the rds file in this directory.

naio_10_cp1700 Symmetric input-output table at basic prices (product by product)

naio_10_pyp1700 Symmetric input-output table at basic prices (product by product) (previous years prices)

naio_10_cp1750 Symmetric input-output table at basic prices (industry by industry)

naio_10_pyp1750 Symmetric input-output table at basic prices (industry by industry) (previous years prices)

naio_10_cp1620 Table of trade and transport margins at basic prices

naio_10_pyp1620 Table of trade and transport margins at previous years' prices

naio_10_cp1630 Table of taxes less subsidies on products at basic prices

naio_10_pyp1630Table of taxes less subsidies on products at previous years' prices

iotables_download(source = "naio_10_cp1700", data_directory = NULL,
  force_download = TRUE)

Arguments

source
See the available list of sources above in the Description.

data_directory
Defaults to NULL, if a valid directory, it will try to save the pre-processed data file here with labelling.

force_download
Defaults to TRUE. If FALSE it will use the existing downloaded file in the data_directory or the temporary directory, if it exists.

Examples

# NOT RUN {
 io_tables <- iotables_download ( source = "naio_10_cp1700" )

# }