Simple input-output table for the Netherlands, 2006.

This simplified SIOT is taken from the Science Policy Integration for Coastal Systems Assessment project's input-output multiplier specification sheet. It is used as a simple example SIOT for controlled analytical results. The column names were slightly altered to resemble more the current Eurostat conventions and the main example dataset germany_1990.

Usage

data(netherlands_2006)

Format

A data frame with 14 observations and 13 variables.

A data frame of 13 observations in 14 variables.

prod_na

Product name, simplified, following the Eurostat conventions

agriculture_group

Simple aggregated agricultural products

mining_group

Simple aggregated mining products

manufacturing_group

Simple aggregated manufacturing products

construction_group

Construction

utilities_group

Simple aggregated utilities products/services

services_group

Simple aggregated services products

TOTAL

Column / row sums, simple summary, not included in the original source

final_consumption_private

Simple aggregated final private use

final_consumption_households

Simple aggregated final household consumption

final_consumption_government

Simple aggregated final government consumption

gross_fixed_capital_formation

Gross fixed capital formation 'GFCF'

exports

Simple aggregated exports

total_use

Simple aggregated total use

Source

Source: Input-Output Multipliers Specification Sheet and Supporting Material in the Spicosa Project Report

See also