Create an output coefficient matrix

Create an output coefficient matrix from the input flow matrix or a symmetric input-output table. If there are zero values in present, they will be changed to 0.000001 and you will get a warning. Some analytical equations cannot be solved with zero elements. You either have faulty input data, or you have to use some sort of data modification to carry on your analysis.

output_coefficient_matrix_create(io_table, total = "tfu", digits = NULL)

Arguments

io_table
A symmetric input-output table or use table created with the iotable_get function which contains the 'total' column. In case you use type="tfu" you need to input a full iotable, create by the iotable_get, because you will need the final demand column.

total
The output='total' (or CPA_TOTAL, depending on the names in your table, default) returns the output coefficients for products (intermediates) while the final_demand returns output coefficients for final demand. See Eurostat Manual, p495 and p507.

digits
An integer showing the precision of the technology matrix in digits. Default is NULL when no rounding is applied.

Value

An output coefficient matrix of data.frame class. The column names are ordered, and the row names are in the first, auxiliary metadata column.

Examples

io_table <- iotable_get () 

output_coefficient_matrix_create ( io_table  = io_table, 
                                   total = 'tfu',
                                   digits = 4 )

#>              iotables_row agriculture_group industry_group construction
#> 1       agriculture_group            0.0258         0.5803       0.0000
#> 2          industry_group            0.0073         0.2822       0.0594
#> 3            construction            0.0017         0.0299       0.0158
#> 4             trade_group            0.0066         0.1346       0.0263
#> 5 business_services_group            0.0053         0.1388       0.0448
#> 6    other_services_group            0.0030         0.0294       0.0034
#>   trade_group business_services_group other_services_group
#> 1      0.0138                  0.0162               0.0174
#> 2      0.0381                  0.0111               0.0281
#> 3      0.0216                  0.0955               0.0373
#> 4      0.1378                  0.0201               0.0389
#> 5      0.0950                  0.2790               0.0494
#> 6      0.0221                  0.0296               0.0434